rsvsvc/requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
opentelemetry-api==0.10b0
opentelemetry-ext-flask==0.10b0
opentelemetry-ext-jaeger==0.10b0
opentelemetry-ext-requests==0.10b0
opentelemetry-ext-wsgi==0.10b0
opentelemetry-instrumentation==0.10b0
opentelemetry-sdk==0.10b0
requests==2.24.0
six==1.15.0
thrift==0.13.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1